refactor(stateupgrader): match with repo struct (#1040)
