+++ internal/schemautil/userconfig/stateupgrader/upgraders.go
@@ -1,15 +1,25 @@
 package stateupgrader
 
 import (
-	v0 "github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/cassandra"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/flink"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/grafana"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/influxdb"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/kafka"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/m3"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/mysql"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/opensearch"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/pg"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/redis"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/serviceintegration"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
 func Cassandra() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceCassandra().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceCassandraStateUpgrade,
+			Type:    cassandra.ResourceCassandra().CoreConfigSchema().ImpliedType(),
+			Upgrade: cassandra.ResourceCassandraStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -18,8 +28,8 @@ func Cassandra() []schema.StateUpgrader {
 func Flink() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceFlink().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceFlinkStateUpgrade,
+			Type:    flink.ResourceFlink().CoreConfigSchema().ImpliedType(),
+			Upgrade: flink.ResourceFlinkStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -28,8 +38,8 @@ func Flink() []schema.StateUpgrader {
 func Grafana() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceGrafana().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceGrafanaStateUpgrade,
+			Type:    grafana.ResourceGrafana().CoreConfigSchema().ImpliedType(),
+			Upgrade: grafana.ResourceGrafanaStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -38,8 +48,8 @@ func Grafana() []schema.StateUpgrader {
 func InfluxDB() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceInfluxDB().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceInfluxDBStateUpgrade,
+			Type:    influxdb.ResourceInfluxDB().CoreConfigSchema().ImpliedType(),
+			Upgrade: influxdb.ResourceInfluxDBStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -48,8 +58,8 @@ func InfluxDB() []schema.StateUpgrader {
 func Kafka() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceKafka().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceKafkaStateUpgrade,
+			Type:    kafka.ResourceKafka().CoreConfigSchema().ImpliedType(),
+			Upgrade: kafka.ResourceKafkaStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -58,8 +68,8 @@ func Kafka() []schema.StateUpgrader {
 func KafkaConnect() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceKafkaConnect().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceKafkaConnectStateUpgrade,
+			Type:    kafka.ResourceKafkaConnect().CoreConfigSchema().ImpliedType(),
+			Upgrade: kafka.ResourceKafkaConnectStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -68,8 +78,8 @@ func KafkaConnect() []schema.StateUpgrader {
 func KafkaMirrormaker() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceKafkaMirrormaker().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceKafkaMirrormakerStateUpgrade,
+			Type:    kafka.ResourceKafkaMirrormaker().CoreConfigSchema().ImpliedType(),
+			Upgrade: kafka.ResourceKafkaMirrormakerStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -78,8 +88,8 @@ func KafkaMirrormaker() []schema.StateUpgrader {
 func M3Aggregator() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceM3Aggregator().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceM3AggregatorStateUpgrade,
+			Type:    m3.ResourceM3Aggregator().CoreConfigSchema().ImpliedType(),
+			Upgrade: m3.ResourceM3AggregatorStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -88,8 +98,8 @@ func M3Aggregator() []schema.StateUpgrader {
 func M3DB() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceM3DBResource().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceM3DBStateUpgrade,
+			Type:    m3.ResourceM3DBResource().CoreConfigSchema().ImpliedType(),
+			Upgrade: m3.ResourceM3DBStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -98,8 +108,8 @@ func M3DB() []schema.StateUpgrader {
 func MySQL() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceMySQLResource().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceMySQLStateUpgrade,
+			Type:    mysql.ResourceMySQLResource().CoreConfigSchema().ImpliedType(),
+			Upgrade: mysql.ResourceMySQLStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -108,8 +118,8 @@ func MySQL() []schema.StateUpgrader {
 func Opensearch() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceOpensearch().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceOpensearchStateUpgrade,
+			Type:    opensearch.ResourceOpensearch().CoreConfigSchema().ImpliedType(),
+			Upgrade: opensearch.ResourceOpensearchStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -118,8 +128,8 @@ func Opensearch() []schema.StateUpgrader {
 func PG() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourcePG().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourcePGStateUpgrade,
+			Type:    pg.ResourcePG().CoreConfigSchema().ImpliedType(),
+			Upgrade: pg.ResourcePGStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -128,8 +138,8 @@ func PG() []schema.StateUpgrader {
 func Redis() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceRedis().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceRedisStateUpgrade,
+			Type:    redis.ResourceRedis().CoreConfigSchema().ImpliedType(),
+			Upgrade: redis.ResourceRedisStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -138,8 +148,8 @@ func Redis() []schema.StateUpgrader {
 func ServiceIntegration() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceServiceIntegration().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceServiceIntegrationStateUpgrade,
+			Type:    serviceintegration.ResourceServiceIntegration().CoreConfigSchema().ImpliedType(),
+			Upgrade: serviceintegration.ResourceServiceIntegrationStateUpgrade,
 			Version: 0,
 		},
 	}
@@ -148,8 +158,8 @@ func ServiceIntegration() []schema.StateUpgrader {
 func ServiceIntegrationEndpoint() []schema.StateUpgrader {
 	return []schema.StateUpgrader{
 		{
-			Type:    v0.ResourceServiceIntegrationEndpoint().CoreConfigSchema().ImpliedType(),
-			Upgrade: v0.ResourceServiceIntegrationEndpointStateUpgrade,
+			Type:    serviceintegration.ResourceServiceIntegrationEndpoint().CoreConfigSchema().ImpliedType(),
+			Upgrade: serviceintegration.ResourceServiceIntegrationEndpointStateUpgrade,
 			Version: 0,
 		},
 	}
