+++ internal/schemautil/userconfig/stateupgrader/v0/grafana/grafana.go
@@ -1,4 +1,4 @@
-package v0
+package grafana
 
 import (
 	"context"
