+++ internal/schemautil/userconfig/stateupgrader/v0/influxdb/influxdb.go
@@ -1,4 +1,4 @@
-package v0
+package influxdb
 
 import (
 	"context"
@@ -12,30 +12,36 @@ import (
 	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/dist"
 )
 
-func aivenM3DBSchema() map[string]*schema.Schema {
-	schemaM3 := schemautil.ServiceCommonSchema()
-	schemaM3[schemautil.ServiceTypeM3] = &schema.Schema{
+func influxDBSchema() map[string]*schema.Schema {
+	s := schemautil.ServiceCommonSchema()
+	s[schemautil.ServiceTypeInfluxDB] = &schema.Schema{
 		Type:        schema.TypeList,
 		Computed:    true,
-		Description: "M3 specific server provided values",
+		Description: "InfluxDB server provided values",
 		Elem: &schema.Resource{
-			Schema: map[string]*schema.Schema{},
+			Schema: map[string]*schema.Schema{
+				"database_name": {
+					Type:        schema.TypeString,
+					Computed:    true,
+					Description: "Name of the default InfluxDB database",
+				},
+			},
 		},
 	}
-	schemaM3[schemautil.ServiceTypeM3+"_user_config"] = dist.ServiceTypeM3db()
+	s[schemautil.ServiceTypeInfluxDB+"_user_config"] = dist.ServiceTypeInfluxdb()
 
-	return schemaM3
+	return s
 }
 
-func ResourceM3DBResource() *schema.Resource {
+func ResourceInfluxDB() *schema.Resource {
 	return &schema.Resource{
-		Description:   "The M3 DB resource allows the creation and management of Aiven M3 services.",
-		CreateContext: schemautil.ResourceServiceCreateWrapper(schemautil.ServiceTypeM3),
+		Description:   "The InfluxDB resource allows the creation and management of Aiven InfluxDB services.",
+		CreateContext: schemautil.ResourceServiceCreateWrapper(schemautil.ServiceTypeInfluxDB),
 		ReadContext:   schemautil.ResourceServiceRead,
 		UpdateContext: schemautil.ResourceServiceUpdate,
 		DeleteContext: schemautil.ResourceServiceDelete,
 		CustomizeDiff: customdiff.Sequence(
-			schemautil.SetServiceTypeIfEmpty(schemautil.ServiceTypeM3),
+			schemautil.SetServiceTypeIfEmpty(schemautil.ServiceTypeInfluxDB),
 			schemautil.CustomizeDiffDisallowMultipleManyToOneKeys,
 			customdiff.IfValueChange("tag",
 				schemautil.TagsShouldNotBeEmpty,
@@ -67,16 +73,16 @@ func ResourceM3DBResource() *schema.Resource {
 			Delete: schema.DefaultTimeout(20 * time.Minute),
 		},
 
-		Schema: aivenM3DBSchema(),
+		Schema: influxDBSchema(),
 	}
 }
 
-func ResourceM3DBStateUpgrade(
+func ResourceInfluxDBStateUpgrade(
 	_ context.Context,
 	rawState map[string]interface{},
 	_ interface{},
 ) (map[string]interface{}, error) {
-	userConfigSlice, ok := rawState["m3db_user_config"].([]interface{})
+	userConfigSlice, ok := rawState["influxdb_user_config"].([]interface{})
 	if !ok {
 		return rawState, nil
 	}
@@ -91,50 +97,26 @@ func ResourceM3DBStateUpgrade(
 	}
 
 	err := typeupgrader.Map(userConfig, map[string]string{
-		"m3coordinator_enable_graphite_carbon_ingest": "bool",
 		"static_ips": "bool",
 	})
 	if err != nil {
 		return rawState, err
 	}
 
-	limitsSlice, ok := userConfig["limits"].([]interface{})
-	if ok && len(limitsSlice) > 0 {
-		limits, ok := limitsSlice[0].(map[string]interface{})
+	influxDBSlice, ok := userConfig["influxdb"].([]interface{})
+	if ok && len(influxDBSlice) > 0 {
+		influxDB, ok := influxDBSlice[0].(map[string]interface{})
 		if ok {
-			err := typeupgrader.Map(limits, map[string]string{
-				"max_recently_queried_series_blocks":          "int",
-				"max_recently_queried_series_disk_bytes_read": "int",
-				"query_docs":               "int",
-				"query_require_exhaustive": "bool",
-				"query_series":             "int",
+			err := typeupgrader.Map(influxDB, map[string]string{
+				"log_queries_after":    "int",
+				"max_connection_limit": "int",
+				"max_row_limit":        "int",
+				"max_select_buckets":   "int",
+				"max_select_point":     "int",
+				"query_timeout":        "int",
 			})
 			if err != nil {
 				return rawState, err
-			}
-		}
-	}
-
-	namespacesSlice, ok := userConfig["namespaces"].([]interface{})
-	if ok && len(namespacesSlice) > 0 {
-		for _, v := range namespacesSlice {
-			namespace, ok := v.(map[string]interface{})
-			if !ok {
-				continue
-			}
-
-			optionsSlice, ok := namespace["options"].([]interface{})
-			if ok && len(optionsSlice) > 0 {
-				options, ok := optionsSlice[0].(map[string]interface{})
-				if ok {
-					err := typeupgrader.Map(options, map[string]string{
-						"snapshot_enabled":    "bool",
-						"writes_to_commitlog": "bool",
-					})
-					if err != nil {
-						return rawState, err
-					}
-				}
 			}
 		}
 	}
@@ -144,7 +126,20 @@ func ResourceM3DBStateUpgrade(
 		privateAccess, ok := privateAccessSlice[0].(map[string]interface{})
 		if ok {
 			err = typeupgrader.Map(privateAccess, map[string]string{
-				"m3coordinator": "bool",
+				"influxdb": "bool",
+			})
+			if err != nil {
+				return rawState, err
+			}
+		}
+	}
+
+	privateLinkAccessSlice, ok := userConfig["privatelink_access"].([]interface{})
+	if ok && len(privateLinkAccessSlice) > 0 {
+		privateLinkAccess, ok := privateLinkAccessSlice[0].(map[string]interface{})
+		if ok {
+			err := typeupgrader.Map(privateLinkAccess, map[string]string{
+				"influxdb": "bool",
 			})
 			if err != nil {
 				return rawState, err
@@ -157,33 +152,10 @@ func ResourceM3DBStateUpgrade(
 		publicAccess, ok := publicAccessSlice[0].(map[string]interface{})
 		if ok {
 			err := typeupgrader.Map(publicAccess, map[string]string{
-				"m3coordinator": "bool",
+				"influxdb": "bool",
 			})
 			if err != nil {
 				return rawState, err
-			}
-		}
-	}
-
-	rulesSlice, ok := userConfig["rules"].([]interface{})
-	if ok && len(rulesSlice) > 0 {
-		rules, ok := rulesSlice[0].(map[string]interface{})
-		if ok {
-			mappingSlice, ok := rules["mapping"].([]interface{})
-			if ok && len(mappingSlice) > 0 {
-				for _, v := range mappingSlice {
-					mapping, ok := v.(map[string]interface{})
-					if !ok {
-						continue
-					}
-
-					err := typeupgrader.Map(mapping, map[string]string{
-						"drop": "bool",
-					})
-					if err != nil {
-						return rawState, err
-					}
-				}
 			}
 		}
 	}
