+++ internal/schemautil/userconfig/stateupgrader/v0/kafka/kafka.go
@@ -1,4 +1,4 @@
-package v0
+package kafka
 
 import (
 	"context"
