+++ internal/schemautil/userconfig/stateupgrader/v0/mysql/mysql.go
@@ -1,4 +1,4 @@
-package v0
+package mysql
 
 import (
 	"context"
@@ -12,36 +12,30 @@ import (
 	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader/v0/dist"
 )
 
-func influxDBSchema() map[string]*schema.Schema {
-	s := schemautil.ServiceCommonSchema()
-	s[schemautil.ServiceTypeInfluxDB] = &schema.Schema{
+func aivenMySQLSchema() map[string]*schema.Schema {
+	schemaMySQL := schemautil.ServiceCommonSchema()
+	schemaMySQL[schemautil.ServiceTypeMySQL] = &schema.Schema{
 		Type:        schema.TypeList,
 		Computed:    true,
-		Description: "InfluxDB server provided values",
+		Description: "MySQL specific server provided values",
 		Elem: &schema.Resource{
-			Schema: map[string]*schema.Schema{
-				"database_name": {
-					Type:        schema.TypeString,
-					Computed:    true,
-					Description: "Name of the default InfluxDB database",
-				},
-			},
+			Schema: map[string]*schema.Schema{},
 		},
 	}
-	s[schemautil.ServiceTypeInfluxDB+"_user_config"] = dist.ServiceTypeInfluxdb()
+	schemaMySQL[schemautil.ServiceTypeMySQL+"_user_config"] = dist.ServiceTypeMysql()
 
-	return s
+	return schemaMySQL
 }
 
-func ResourceInfluxDB() *schema.Resource {
+func ResourceMySQLResource() *schema.Resource {
 	return &schema.Resource{
-		Description:   "The InfluxDB resource allows the creation and management of Aiven InfluxDB services.",
-		CreateContext: schemautil.ResourceServiceCreateWrapper(schemautil.ServiceTypeInfluxDB),
+		Description:   "The MySQL resource allows the creation and management of Aiven MySQL services.",
+		CreateContext: schemautil.ResourceServiceCreateWrapper(schemautil.ServiceTypeMySQL),
 		ReadContext:   schemautil.ResourceServiceRead,
 		UpdateContext: schemautil.ResourceServiceUpdate,
 		DeleteContext: schemautil.ResourceServiceDelete,
 		CustomizeDiff: customdiff.Sequence(
-			schemautil.SetServiceTypeIfEmpty(schemautil.ServiceTypeInfluxDB),
+			schemautil.SetServiceTypeIfEmpty(schemautil.ServiceTypeMySQL),
 			schemautil.CustomizeDiffDisallowMultipleManyToOneKeys,
 			customdiff.IfValueChange("tag",
 				schemautil.TagsShouldNotBeEmpty,
@@ -73,16 +67,16 @@ func ResourceInfluxDB() *schema.Resource {
 			Delete: schema.DefaultTimeout(20 * time.Minute),
 		},
 
-		Schema: influxDBSchema(),
+		Schema: aivenMySQLSchema(),
 	}
 }
 
-func ResourceInfluxDBStateUpgrade(
+func ResourceMySQLStateUpgrade(
 	_ context.Context,
 	rawState map[string]interface{},
 	_ interface{},
 ) (map[string]interface{}, error) {
-	userConfigSlice, ok := rawState["influxdb_user_config"].([]interface{})
+	userConfigSlice, ok := rawState["mysql_user_config"].([]interface{})
 	if !ok {
 		return rawState, nil
 	}
@@ -97,23 +91,60 @@ func ResourceInfluxDBStateUpgrade(
 	}
 
 	err := typeupgrader.Map(userConfig, map[string]string{
-		"static_ips": "bool",
+		"backup_hour":             "int",
+		"backup_minute":           "int",
+		"binlog_retention_period": "int",
+		"static_ips":              "bool",
 	})
 	if err != nil {
 		return rawState, err
 	}
 
-	influxDBSlice, ok := userConfig["influxdb"].([]interface{})
-	if ok && len(influxDBSlice) > 0 {
-		influxDB, ok := influxDBSlice[0].(map[string]interface{})
+	migrationSlice, ok := userConfig["migration"].([]interface{})
+	if ok && len(migrationSlice) > 0 {
+		migration, ok := migrationSlice[0].(map[string]interface{})
 		if ok {
-			err := typeupgrader.Map(influxDB, map[string]string{
-				"log_queries_after":    "int",
-				"max_connection_limit": "int",
-				"max_row_limit":        "int",
-				"max_select_buckets":   "int",
-				"max_select_point":     "int",
-				"query_timeout":        "int",
+			err = typeupgrader.Map(migration, map[string]string{
+				"port": "int",
+				"ssl":  "bool",
+			})
+			if err != nil {
+				return rawState, err
+			}
+		}
+	}
+
+	mysqlSlice, ok := userConfig["mysql"].([]interface{})
+	if ok && len(mysqlSlice) > 0 {
+		mysql, ok := mysqlSlice[0].(map[string]interface{})
+		if ok {
+			err = typeupgrader.Map(mysql, map[string]string{
+				"connect_timeout":                  "int",
+				"group_concat_max_len":             "int",
+				"information_schema_stats_expiry":  "int",
+				"innodb_change_buffer_max_size":    "int",
+				"innodb_flush_neighbors":           "int",
+				"innodb_ft_min_token_size":         "int",
+				"innodb_lock_wait_timeout":         "int",
+				"innodb_log_buffer_size":           "int",
+				"innodb_online_alter_log_max_size": "int",
+				"innodb_print_all_deadlocks":       "bool",
+				"innodb_read_io_threads":           "int",
+				"innodb_rollback_on_timeout":       "bool",
+				"innodb_thread_concurrency":        "int",
+				"innodb_write_io_threads":          "int",
+				"interactive_timeout":              "int",
+				"long_query_time":                  "float",
+				"max_allowed_packet":               "int",
+				"max_heap_table_size":              "int",
+				"net_buffer_length":                "int",
+				"net_read_timeout":                 "int",
+				"net_write_timeout":                "int",
+				"slow_query_log":                   "bool",
+				"sort_buffer_size":                 "int",
+				"sql_require_primary_key":          "bool",
+				"tmp_table_size":                   "int",
+				"wait_timeout":                     "int",
 			})
 			if err != nil {
 				return rawState, err
@@ -126,7 +157,9 @@ func ResourceInfluxDBStateUpgrade(
 		privateAccess, ok := privateAccessSlice[0].(map[string]interface{})
 		if ok {
 			err = typeupgrader.Map(privateAccess, map[string]string{
-				"influxdb": "bool",
+				"mysql":      "bool",
+				"mysqlx":     "bool",
+				"prometheus": "bool",
 			})
 			if err != nil {
 				return rawState, err
@@ -139,7 +172,9 @@ func ResourceInfluxDBStateUpgrade(
 		privateLinkAccess, ok := privateLinkAccessSlice[0].(map[string]interface{})
 		if ok {
 			err := typeupgrader.Map(privateLinkAccess, map[string]string{
-				"influxdb": "bool",
+				"mysql":      "bool",
+				"mysqlx":     "bool",
+				"prometheus": "bool",
 			})
 			if err != nil {
 				return rawState, err
@@ -152,7 +187,9 @@ func ResourceInfluxDBStateUpgrade(
 		publicAccess, ok := publicAccessSlice[0].(map[string]interface{})
 		if ok {
 			err := typeupgrader.Map(publicAccess, map[string]string{
-				"influxdb": "bool",
+				"mysql":      "bool",
+				"mysqlx":     "bool",
+				"prometheus": "bool",
 			})
 			if err != nil {
 				return rawState, err
