+++ internal/schemautil/userconfig/stateupgrader/v0/opensearch/opensearch.go
@@ -1,4 +1,4 @@
-package v0
+package opensearch
 
 import (
 	"context"
