+++ internal/schemautil/userconfig/stateupgrader/v0/pg/pg.go
@@ -1,4 +1,4 @@
-package v0
+package pg
 
 import (
 	"context"
