+++ internal/schemautil/userconfig/stateupgrader/v0/serviceintegration/service_integration_endpoint.go
@@ -1,4 +1,4 @@
-package v0
+package serviceintegration
 
 import (
 	"context"
